[refactor] khushal87 Card
