+++ src/Card/__tests__/Card.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import toJson from 'enzyme-to-json';
 import theme from '../../config/theme';
-import Card from '../Card';
+import Card from '../index';
 
 describe('Card Component', () => {
   it('should render without issues', () => {
